requirements.txt
PyMuPDF==1.24.14
pillow==11.0.0
pyodbc==5.2.0
python-dotenv==1.0.1
tkcalendar==1.6.1